app/requirements.txt
Flask==2.1.2
Flask_SQLAlchemy==2.5.1
python-dotenv==0.20.0
SQLAlchemy==1.4.36
psycopg2